->20.31

+++ src/com/andrew/command/admin/application/AdminDeleteApplicationCommand.java
@@ -1,9 +1,11 @@
 package com.andrew.command.admin.application;
 
 import com.andrew.action.Action;
+import com.andrew.action.ForwardAction;
 import com.andrew.action.RedirectAction;
 import com.andrew.command.Command;
 import com.andrew.dao.ApplicationDao;
+import com.andrew.util.SessionChecker;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
@@ -11,6 +13,7 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 import static com.andrew.constant.CommandPathConstant.ADMIN_ALL_APPLICATIONS_COMMAND;
+import static com.andrew.constant.JspPathConstant.LOG_IN_ADMIN_JSP;
 
 /**
  * Created by Andrew on 12.04.2017.
@@ -25,8 +28,12 @@ public class AdminDeleteApplicationCommand implements Command {
 
     @Override
     public Action execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
-        int applicationId = Integer.parseInt(request.getParameter("id"));
-        dao.deleteApplication(applicationId);
-        return new RedirectAction(ADMIN_ALL_APPLICATIONS_COMMAND);
+        if(SessionChecker.adminInSession(request)) {
+            int applicationId = Integer.parseInt(request.getParameter("id"));
+            dao.deleteApplication(applicationId);
+            return new RedirectAction(ADMIN_ALL_APPLICATIONS_COMMAND);
+        } else {
+            return new ForwardAction(LOG_IN_ADMIN_JSP);
+        }
     }
 }
